capture_reconnaissance/requirements.txt
opencv-python-headless==4.9.0.80
ultralytics==8.1.0
mysql-connector-python==8.3.0
numpy==1.26.4
PyYAML==6.0.1
python-dotenv==1.0.1
apache-airflow==2.8.1
Pillow==10.2.0
